Rename group vars file (etke_servers -> etke_overrides)
The group hierrarchy in the inventory has been redone,
so that we can do overrides correctly.

Until now we had 2 levels (`etke_servers` -> `matrix_servers`)
and we expected that `etke_servers` was taking precedence.
That's not the case though:
https://docs.ansible.com/ansible/latest/inventory_guide/intro_inventory.html#how-variables-are-merged

From now on, we have `etke_servers` as a common parent,
then `matrix_servers` and `etke_overrides` as children of it.
The `etke_overrides` child has a higher group priority (among its
siblings, like `matrix_servers`), so that it can do overrides.